Working loading logic
Added player to world_root_no_touch group so the world root won't kill the player node and not load in when the game gets loaded.
Added nodepath to player script as export variable to world root so the world root can find the player node.
Added closing of pause menu when game gets loaded so the UI doesn't break.